feat(metadata-sidebar): PR comments
